Move character gender function to utils file
